started the creation of dedicated tweet object. Will allow for moving tweets, and reduce client workload (clients only render tweets they use)
